+++ src/backend/joanie/tests/core/utils/test_issuers_certificate_document.py
@@ -93,9 +93,10 @@ class UtilsIssuersCertificateGenerateDocumentTestCase(TestCase):
         )
 
         document_text = pdf_extract_text(BytesIO(document)).replace("\n", "")
-        self.assertRegex(document_text, "Certificate")
-        self.assertRegex(
-            document_text, f"Issued on {certificate.issued_on.strftime('%m/%d/%Y')}"
+        self.assertRegex(document_text, "CERTIFICATE")
+        self.assertIn(
+            " ".join(f"Issued on {certificate.issued_on.strftime('%m/%d/%Y')}".upper()),
+            document_text,
         )
         self.assertRegex(document_text, r"Richie Cunningham.*Graded product")
         self.assertRegex(